+++ packages/isar_generator/test/errors/index/object_hashed.dart
@@ -1,0 +1,14 @@
+// objects may not be indexed
+
+import 'package:isar/isar.dart';
+
+@collection
+class Model {
+  Id? id;
+
+  @Index()
+  EmbeddedModel? obj;
+}
+
+@embedded
+class EmbeddedModel {}